+++ src/pages/blind-match/blind-match.tsx
@@ -1,103 +1,34 @@
-import { useState } from 'react';
+import { useParams } from 'react-router-dom';
 
-import AgreementOption from '@shared/components/agreement-option/agreement-option.tsx';
-import Button from '@shared/components/button/button';
-import Title from '@shared/components/title/title';
-import { IcSvgCaution } from '@shared/icons';
+import Tab from '@shared/components/tab/tab';
 
-import * as styles from './blind-match.css.ts';
-import ConfirmModal from './components/modal/confirm-modal.tsx';
-import UseInfoForm from './components/use-info-form/use-info-form';
-import { BLIND_MATCH_TEXT } from './constants/blind-match-text';
+import EntryPage from './components/entrypage/entry';
 
 const BlindMatch = () => {
-  const [agreed, setAgreed] = useState(false);
-  const [instaId, setInstaId] = useState('');
-  const [phoneNumber, setPhoneNumber] = useState('');
-  const [name, setName] = useState('');
-  const [gender, setGender] = useState('');
-  const [isModalOpen, setIsModalOpen] = useState(false);
+  const { day } = useParams<{ day: string }>();
 
-  const handleConsentChange = (isAgreed: boolean) => {
-    setAgreed(isAgreed);
-  };
+  const initialDay = day || '1';
 
-  const handleApplyClick = () => {
-    const phoneRegex = /^\d{2,3}-\d{4}-\d{4}$/;
-    if (name && instaId && phoneNumber && gender && agreed) {
-      if (phoneRegex.test(phoneNumber)) {
-        setIsModalOpen(true);
-      }
-    }
-  };
-
-  const handleConfirm = () => {
-    setIsModalOpen(false);
-  };
-
-  const handleCloseModal = () => {
-    setIsModalOpen(false);
-  };
   return (
     <>
-      <div className={styles.title}>
-        <Title
-          mainTitle={BLIND_MATCH_TEXT.TITLE}
-          subTitle={BLIND_MATCH_TEXT.SUBTITLE}
-        />
-      </div>
-
-      <div className={styles.container}>
-        <div className={styles.textcontainer}>
-          <p className={styles.message}>{BLIND_MATCH_TEXT.MESSAGE}</p>
-          <p className={styles.time}>{BLIND_MATCH_TEXT.TIME}</p>
-        </div>
-
-        <UseInfoForm
-          name={name}
-          instaId={instaId}
-          phoneNumber={phoneNumber}
-          gender={gender}
-          onNameChange={(e) => setName(e.target.value)}
-          onInstaIdChange={(e) => setInstaId(e.target.value)}
-          onPhoneNumberChange={(value) => setPhoneNumber(value)}
-          onGenderChange={(value) => setGender(value)}
-        />
-
-        <div className={styles.attention}>
-          <IcSvgCaution
-            width={12}
-            height={12}
-          />
-          {BLIND_MATCH_TEXT.ATTENTION}
-        </div>
-
-        <div className={styles.agreementoption}>
-          <AgreementOption
-            label={BLIND_MATCH_TEXT.CONSENT}
-            navigateTo=""
-            checked={agreed}
-            onChange={handleConsentChange}
-          />
-        </div>
-
-        <Button onClick={handleApplyClick}>번호팅 신청하기</Button>
-
-        <div className={styles.notice}>
-          <IcSvgCaution
-            width={12}
-            height={12}
-          />
-          {BLIND_MATCH_TEXT.NOTICE}
-        </div>
-      </div>
-      <ConfirmModal
-        isModalOpen={isModalOpen}
-        instaId={instaId}
-        phoneNumber={phoneNumber}
-        onClose={handleCloseModal}
-        onConfirm={handleConfirm}
-      />
+      <Tab.Container initialValue={`day-${initialDay}`}>
+        <Tab.List>
+          <Tab.Item value="day-1">1일차 매칭</Tab.Item>
+          <Tab.Item value="day-2">2일차 매칭</Tab.Item>
+          <Tab.Item value="day-3">3일차 매칭</Tab.Item>
+        </Tab.List>
+        <Tab.Panels>
+          <Tab.Panel value="day-1">
+            <EntryPage currentDay="1일차 매칭" />
+          </Tab.Panel>
+          <Tab.Panel value="day-2">
+            <EntryPage currentDay="2일차 매칭" />
+          </Tab.Panel>
+          <Tab.Panel value="day-3">
+            <EntryPage currentDay="3일차 매칭" />
+          </Tab.Panel>
+        </Tab.Panels>
+      </Tab.Container>
     </>
   );
 };
