+++ src/pages/blind-match/components/entrypage/entry.css.ts
