+++ src/pages/blind-match/constants/blind-match-text.ts
@@ -1,7 +1,7 @@
 export const BLIND_MATCH_TEXT = {
   TITLE: '번호팅',
   SUBTITLE: '랜덤 매칭으로 새로운 인연을 만들어주는 서비스',
-  MESSAGE: '1일차 매칭에 아직 신청하지 않았습니다!',
+  MESSAGE: ' 매칭에 아직 신청하지 않았습니다!',
   TIME: '[신청 가능시간 00:00 ~ 17:30]',
   ATTENTION: '모든 정보를 입력해주세요.',
   CONSENT: '정보 제공 동의하기(필수)',
